Add payment by turanline details

+++ backend/cart/tasks.py
@@ -1,12 +1,14 @@
 from celery import shared_task
 
-from . import models
+from . import enums, models
 
 
 @shared_task
 def delete_order_if_not_payed(order_id: int) -> None:
-    order = models.Order.objects.get(
-        id=order_id
-    )
-    if not order.is_paid:
+    order = models.Order.objects.get(id=order_id)
+
+    if (
+        not order.is_paid
+        and order.status != enums.OrderStatuses.ON_PAYMENT
+    ):
         order.delete()
